fix(omniauth): user with linked Google account can't sign in
This happened because the `User#self.from_omniauth` method returned `nil` if an account with the chosen email was found. However, it didn't check if the existing account matched the `auth.provider` and `auth.uid` provided, thus preventing the user from signing in. This commit fixes this bug and adds comments to the `Users::OmniauthCallbacksController#google_oauth2` method for better documentation.